SwapyExchange tests
